+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added experimental support for image generation with OpenAI DALL-E models - see `?aiimage` for more details.
 
 ### Fixed
+
+## [0.15.0]
+
+### Added
+- Added experimental support for image generation with OpenAI DALL-E models, eg, `msg = aiimage("A white cat on a car")`. See `?aiimage` for more details.
 
 ## [0.14.0]